% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd2shiny.R
\name{knit_rmd_to_shiny}
\alias{knit_rmd_to_shiny}
\title{A shiny runtime for parametrized r markdown reports.}
\usage{
knit_rmd_to_shiny(file = NULL, params = NULL, shiny_args = NULL,
  save_caption = "Save", run_caption = "Run Report",
  encoding = "UTF-8", defaultlogo = "UCCH_Front.png")
}
\arguments{
\item{file}{Path to the R Markdown document with configurable parameters.}

\item{params}{A named list of optional parameter overrides used in place of the document defaults.}

\item{shiny_args}{Additional arguments to \code{\link[shiny:runApp]{runApp}}.}

\item{save_caption}{Caption to use use for button that saves/confirms parameters.}

\item{run_caption}{Caption of run button}

\item{encoding}{The encoding of the input file; see \code{\link{file}}.}
}
\value{
named list with overridden parameter names and value.
}
\description{
The knit_rmd_to_shiny function starts a shiny app that reads the YAML header of a given 
rmd document and shows a corresponding input form for the parameters defined in the yaml
header. knit_rmd_to_shiny is a modified version of the knit_params_ask function from the rmarkdown
package that is used fpor parametrized rmd reports in RStudio. 

In addition to that special YAML tags can be used to modify the appearance of the shiny app
}
